# perf: Put function calls last in example configurations.