--- a/examples/abac_rule_model.conf
+++ b/examples/abac_rule_model.conf
@@ -8,4 +8,4 @@ p = sub_rule, obj, act
 e = some(where (p.eft == allow))
 
 [matchers]
-m = eval(p.sub_rule) && r.obj == p.obj && r.act == p.act
+m = r.obj == p.obj && r.act == p.act && eval(p.sub_rule)