--- a/examples/eval_operator_model.conf
+++ b/examples/eval_operator_model.conf
@@ -8,4 +8,4 @@ p = sub_rule, obj_rule, act
 e = some(where (p.eft == allow))
 
 [matchers]
-m = eval(p.sub_rule) && eval(p.obj_rule) && (p.act == '*' || r.act == p.act)
+m = (p.act == '*' || r.act == p.act) && eval(p.sub_rule) && eval(p.obj_rule)