--- a/examples/glob_model.conf
+++ b/examples/glob_model.conf
@@ -8,4 +8,4 @@ p = sub, obj, act
 e = some(where (p.eft == allow))
 
 [matchers]
-m = r.sub == p.sub && globMatch(r.obj, p.obj) && r.act == p.act
+m = r.sub == p.sub && r.act == p.act && globMatch(r.obj, p.obj)