--- a/examples/ipmatch_model.conf
+++ b/examples/ipmatch_model.conf
@@ -8,4 +8,4 @@ p = sub, obj, act
 e = some(where (p.eft == allow))
 
 [matchers]
-m = ipMatch(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
+m = r.obj == p.obj && r.act == p.act && ipMatch(r.sub, p.sub)