--- a/examples/multiple_policy_definitions_model.conf
+++ b/examples/multiple_policy_definitions_model.conf
@@ -14,6 +14,6 @@ e = some(where (p.eft == allow))
 
 [matchers]
 #RABC
-m = g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
+m = r.obj == p.obj && r.act == p.act && g(r.sub, p.sub)
 #ABAC
-m2 = eval(p2.sub_rule) && r2.obj == p2.obj && r2.act == p2.act
+m2 = r2.obj == p2.obj && r2.act == p2.act && eval(p2.sub_rule)