--- a/examples/priority_model_explicit.conf
+++ b/examples/priority_model_explicit.conf
@@ -11,4 +11,4 @@ g = _, _
 e = priority(p.eft) || deny
 
 [matchers]
-m = g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
+m = r.obj == p.obj && r.act == p.act && g(r.sub, p.sub)