--- a/examples/rbac_model.conf
+++ b/examples/rbac_model.conf
@@ -11,4 +11,4 @@ g = _, _
 e = some(where (p.eft == allow))
 
 [matchers]
-m = g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
+m = r.obj == p.obj && r.act == p.act && g(r.sub, p.sub)