--- a/examples/rbac_model_in_multi_line.conf
+++ b/examples/rbac_model_in_multi_line.conf
@@ -11,5 +11,5 @@ g = _, _
 e = some(where (p.eft == allow))
 
 [matchers]
-m = g(r.sub, p.sub) && r.obj == p.obj \
- && r.act == p.act
+m = r.obj == p.obj && r.act == p.act \
+  && g(r.sub, p.sub)