--- a/examples/rbac_model_matcher_using_in_op_bracket.conf
+++ b/examples/rbac_model_matcher_using_in_op_bracket.conf
@@ -11,4 +11,4 @@ g = _, _
 e = some(where (p.eft == allow))
 
 [matchers]
-m = g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act || r.obj in ['data2', 'data3']
+m = r.obj == p.obj && r.act == p.act && g(r.sub, p.sub) || r.obj in ['data2', 'data3']