--- a/examples/rbac_with_resource_roles_model.conf
+++ b/examples/rbac_with_resource_roles_model.conf
@@ -12,4 +12,4 @@ g2 = _, _
 e = some(where (p.eft == allow))
 
 [matchers]
-m = g(r.sub, p.sub) && g2(r.obj, p.obj) && r.act == p.act
+m = r.act == p.act && g(r.sub, p.sub) && g2(r.obj, p.obj)