--- a/examples/subject_priority_model.conf
+++ b/examples/subject_priority_model.conf
@@ -11,4 +11,4 @@ g = _, _
 e = subjectPriority(p.eft) || deny
 
 [matchers]
-m = g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
+m = r.obj == p.obj && r.act == p.act && g(r.sub, p.sub)