--- a/examples/subject_priority_model_with_domain.conf
+++ b/examples/subject_priority_model_with_domain.conf
@@ -11,4 +11,4 @@ g = _, _, _
 e = subjectPriority(p.eft) || deny
 
 [matchers]
-m = g(r.sub, p.sub, r.dom) && r.dom == p.dom && r.obj == p.obj && r.act == p.act
+m = r.dom == p.dom && r.obj == p.obj && r.act == p.act && g(r.sub, p.sub, r.dom)