String comparisons are a *lot* cheaper than govaluate function calls,
so putting them last in the matcher function can save a shocking amount
of CPU time and allocations, while being functionally identical.

This change reorders the example configurations to put function calls
last where applicable, to provide a good example for users to base
their own configuration off of.

The results of a benchmark where I compared the two options:

    g_first:   g(r.sub, p.sub) && r.obj == p.obj && r.act == p.act
    eq_first:  r.obj == p.obj && r.act == p.act && g(r.sub, p.sub)

    Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkRBACModelSizes$ github.com/casbin/casbin/v2

    goos: linux
    goarch: amd64
    pkg: github.com/casbin/casbin/v2
    cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
    BenchmarkRBACModelSizes/small/g_first-4      14827      79860 ns/op    38767 B/op      874 allocs/op
    BenchmarkRBACModelSizes/small/eq_first-4     28083      42782 ns/op    16938 B/op      395 allocs/op
    BenchmarkRBACModelSizes/medium/g_first-4      1544     774049 ns/op   372522 B/op     8141 allocs/op
    BenchmarkRBACModelSizes/medium/eq_first-4     4285     284674 ns/op   101990 B/op     2572 allocs/op
    BenchmarkRBACModelSizes/large/g_first-4        159    7562438 ns/op  3533126 B/op    80803 allocs/op
    BenchmarkRBACModelSizes/large/eq_first-4       462    2707900 ns/op   943430 B/op    24257 allocs/op
    PASS
    ok   github.com/casbin/casbin/v2 10.320s

This benchmark is currently only local, but I can commit it to the test
suite if desired, so that the difference can be obvious.