@thetorpedodog 2022-03-08
After a few days of this change being in production for us, we’ve found that it’s reduced Casbin latency by about 60%!