@sagilio 2022-03-22
According to the principle of expression, this is a valid operation to improve performance. I think we can add this to the docs.
It sounds good If we can auto optimize the matcher.