@hsluoyz 2022-03-22
@thetorpedodog can we do this shifting automatically in the code? I prefer not changing the order in the model text for better readability.